Removed STR -1.
Spymaster is now equally strong (that is, completely normal) as the Advisor, rather than weaker.

While the combat Hand is +2 and the non-combat Advisor was 0, the 'quiet combat' Spymaster was -1. This should allow -STR stat packs to be the means of creating a weak Spymaster or reduce the pain of a -STR race or age as spymaster. Grabbing, breaking out of grabs and the like are all crucial for 'wet work'.

This commit is isolated and individual specifically to allow feedback and review, since any stat adjustment deserves scrutiny.